begin refactor, add InvoiceProcessor and InvoiceAppGUI
